\name{dbClearResult}
\alias{dbClearResult}
\title{Clears resources associated with a query result}
\description{
  \code{dbClearResult} is a generic function that, when called on a
  result object, clears any resources associated with that object.
}
\usage{
dbClearResult(result)
}
\arguments{
  \item{result}{A query result object.}
}
\references{\url{http://rdbi.sourceforge.net/}}
\author{Timothy H. Keitt}

\seealso{\code{\link{dbSendQuery}}, \code{\link{data.frame}}}

\keyword{methods}
\keyword{data}
